Fix API type conversion
